Replace uses of celery.utils.time
The FixedOffset and timezone imports were actually from pytz, which has
been deprecated and therefore removed as a celery dependency. Instead,
use zoneinfo in Python 3.9+ and backports.zoneinfo in earlier Python 3
versions to look up timezones by name, and use dateutil.timezone.utc
and dateutil.timezone(timedelta(...)) to construct timezones that are a
fixed offset from UTC.

For background and inspiration, see pganssle/pytz-deprecation-shim, but
that is not necessary as a dependency for these uses.

Fixes #244.